WEB-7. `SettingsState` structure was refactored, unnecessary fields were removed, settings reducers file was moved to module directory for consistent structure of project directories.
